# projects/backend/services/telemetry-ingest-service/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Telemetry Ingest Service API
  version: 0.1.0
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  service: { type: string }
                  env: { type: string }
  /api/v1/telemetry:
    post:
      summary: Ingest telemetry batch
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer sensor token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelemetryIngest"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  accepted: { type: integer }
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /api/v1/telemetry/stream:
    get:
      summary: Stream telemetry records (SSE)
      description: >
        Server-Sent Events stream for a single sensor. Requires Bearer token. In auth-proxy setup,
        this endpoint can be accessed with either a sensor token or a user session.
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer sensor token (or user JWT via auth-proxy)
        - name: X-Access-Token
          in: header
          required: false
          schema:
            type: string
          description: Optional token for SSE clients that cannot set Authorization
        - name: access_token
          in: query
          required: false
          schema:
            type: string
          description: Optional token for SSE clients that cannot set Authorization
        - name: sensor_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: since_id
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: max_events
          in: query
          required: false
          schema:
            type: integer
            nullable: true
        - name: idle_timeout_seconds
          in: query
          required: false
          schema:
            type: number
            default: 30
      responses:
        "200":
          description: OK (text/event-stream)
          content:
            text/event-stream:
              schema:
                type: string
  /api/v1/telemetry/query:
    get:
      summary: Query telemetry records for a capture session
      description: >
        Historical query endpoint for telemetry records tied to a capture session.
        Requires user JWT (via auth-proxy). Supports cursor-based pagination by id.
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
          description: Bearer user token (auth-proxy injects from cookies)
        - name: access_token
          in: query
          required: false
          schema:
            type: string
          description: Optional token for clients that cannot set Authorization
        - name: capture_session_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: sensor_id
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
              format: uuid
          style: form
          explode: true
        - name: since_id
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 2000
        - name: include_late
          in: query
          required: false
          schema:
            type: boolean
            default: true
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetryQueryResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
components:
  schemas:
    TelemetryReading:
      type: object
      required: [timestamp, raw_value]
      properties:
        timestamp:
          type: string
          format: date-time
        raw_value:
          type: number
        physical_value:
          type: number
          nullable: true
        meta:
          type: object
          additionalProperties: true
    TelemetryIngest:
      type: object
      required: [sensor_id, readings]
      properties:
        sensor_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
          nullable: true
          description: "Optional. Usually omitted: service will infer run_id from project's active capture session if present."
        capture_session_id:
          type: string
          format: uuid
          nullable: true
          description: "Optional. Usually omitted: service will attach active capture session automatically. If provided but not active, it will not be attached."
        meta:
          type: object
          description: Optional batch-level metadata merged into each reading meta
          additionalProperties: true
        readings:
          type: array
          minItems: 1
          maxItems: 10000
          items:
            $ref: "#/components/schemas/TelemetryReading"
    TelemetryQueryRecord:
      type: object
      required: [id, sensor_id, project_id, timestamp, raw_value, meta]
      properties:
        id:
          type: integer
        project_id:
          type: string
          format: uuid
        sensor_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        raw_value:
          type: number
        physical_value:
          type: number
          nullable: true
        run_id:
          type: string
          format: uuid
          nullable: true
        capture_session_id:
          type: string
          format: uuid
          nullable: true
        meta:
          type: object
          additionalProperties: true
    TelemetryQueryResponse:
      type: object
      required: [points]
      properties:
        points:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryQueryRecord"
        next_since_id:
          type: integer
          nullable: true